+++ lib/mini_utils.dart
@@ -7,3 +7,6 @@ export 'src/widgets/digital_clock.dart' show digitalClock;
 export 'src/widgets/draggable_container.dart' show DraggableContainer;
 export 'src/widgets/line_connector_painter.dart'
     show LineConnectorPainter, Tuple2;
+export 'src/themes/custom_theme_dark.dart' show cdevSuiteDarkTheme;
+export 'src/themes/custom_theme_light.dart' show cdevSuiteLightTheme;
+export 'src/functions/color_inverter.dart' show colorInverter;