+++ lib/src/themes/custom_theme_dark.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+ThemeData cdevSuiteDarkTheme = ThemeData(
+  colorScheme: ColorScheme.fromSeed(
+      seedColor: const Color.fromARGB(255, 35, 94, 113),
+      brightness: Brightness.dark),
+  useMaterial3: true,
+);