+++ lib/src/themes/custom_theme_light.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+ThemeData cdevSuiteLightTheme = ThemeData(
+  colorScheme: ColorScheme.fromSeed(
+      seedColor: const Color.fromARGB(255, 148, 229, 255),
+      brightness: Brightness.light),
+  useMaterial3: true,
+);